[dashboard] Move `components/dashboard/*` to `pages/dashboard/components/*`
