+++ app/gui/src/dashboard/components/MenuEntry.tsx
@@ -1,7 +1,6 @@
 /** @file An entry in a menu. */
 import BlankIcon from '#/assets/blank.svg'
 import * as aria from '#/components/aria'
-import KeyboardShortcut from '#/components/dashboard/KeyboardShortcut'
 import { useDialogContext } from '#/components/Dialog'
 import { Icon } from '#/components/Icon'
 import FocusRing from '#/components/styled/FocusRing'
@@ -10,6 +9,7 @@ import { useVisualTooltip } from '#/components/VisualTooltip'
 import type * as inputBindings from '#/configurations/inputBindings'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useSyncRef } from '#/hooks/syncRefHooks'
+import KeyboardShortcut from '#/pages/dashboard/components/KeyboardShortcut'
 import * as inputBindingsProvider from '#/providers/InputBindingsProvider'
 import { unsetModal } from '#/providers/ModalProvider'
 import * as sanitizedEventTargets from '#/utilities/sanitizedEventTargets'
