+++ app/gui/src/dashboard/layouts/AssetPanel/components/AssetProperties.tsx
@@ -2,9 +2,6 @@
 import PenIcon from '#/assets/pen.svg'
 import { Heading } from '#/components/aria'
 import { Button, CopyButton } from '#/components/Button'
-import SharedWithColumn from '#/components/dashboard/column/SharedWithColumn'
-import { DatalinkFormInput } from '#/components/dashboard/DatalinkInput'
-import Label from '#/components/dashboard/Label'
 import { Form } from '#/components/Form'
 import { ResizableContentEditableInput } from '#/components/Inputs'
 import { Result } from '#/components/Result'
@@ -16,6 +13,9 @@ import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useSpotlight } from '#/hooks/spotlightHooks'
 import { type Category } from '#/layouts/Drive/Categories'
 import { UpsertSecretForm } from '#/modals/UpsertSecretModal'
+import SharedWithColumn from '#/pages/dashboard/components/column/SharedWithColumn'
+import { DatalinkFormInput } from '#/pages/dashboard/components/DatalinkInput'
+import Label from '#/pages/dashboard/components/Label'
 import { useFullUserSession } from '#/providers/AuthProvider'
 import { useFeatureFlags } from '#/providers/FeatureFlagsProvider'
 import type Backend from '#/services/Backend'
