+++ app/gui/src/dashboard/layouts/AssetSearchBar.tsx
@@ -1,7 +1,6 @@
 /** @file A search bar containing a text input, and a list of suggestions. */
 import * as aria from '#/components/aria'
 import { Button } from '#/components/Button'
-import Label from '#/components/dashboard/Label'
 import { DIALOG_BACKGROUND } from '#/components/Dialog'
 import { Icon } from '#/components/Icon'
 import FocusRing from '#/components/styled/FocusRing'
@@ -9,6 +8,7 @@ import { Text } from '#/components/Text'
 import { backendQueryOptions } from '#/hooks/backendHooks'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useSyncRef } from '#/hooks/syncRefHooks'
+import Label from '#/pages/dashboard/components/Label'
 import * as modalProvider from '#/providers/ModalProvider'
 import type Backend from '#/services/Backend'
 import type { Label as BackendLabel } from '#/services/Backend'
