+++ app/gui/src/dashboard/layouts/Drive/__test__/compareAssets.test.ts
@@ -1,6 +1,6 @@
 /** @file Tests for comparing assets. */
-import { Column, type SortableColumn } from '#/components/dashboard/column/columnUtils'
 import { assetCompareFunction } from '#/layouts/Drive/compareAssets'
+import { Column, type SortableColumn } from '#/pages/dashboard/components/column/columnUtils'
 import { SortDirection, type SortInfo } from '#/utilities/sorting'
 import * as fc from '@fast-check/vitest'
 import { DirectoryId, createPlaceholderFileAsset } from 'enso-common/src/services/Backend'
