+++ app/gui/src/dashboard/layouts/Settings/KeyboardShortcutsSettingsSection.tsx
@@ -4,7 +4,6 @@ import CrossIcon from '#/assets/cross.svg'
 import Plus2Icon from '#/assets/plus2.svg'
 import ReloadIcon from '#/assets/reload.svg'
 import { Button } from '#/components/Button'
-import KeyboardShortcut from '#/components/dashboard/KeyboardShortcut'
 import { Dialog } from '#/components/Dialog'
 import { Icon } from '#/components/Icon'
 import { ACTION_TO_TEXT_ID } from '#/components/MenuEntry'
@@ -13,6 +12,7 @@ import type { DashboardBindingKey } from '#/configurations/inputBindings'
 import { useRefresh } from '#/hooks/refreshHooks'
 import CaptureKeyboardShortcutModal from '#/modals/CaptureKeyboardShortcutModal'
 import ConfirmDeleteModal from '#/modals/ConfirmDeleteModal'
+import KeyboardShortcut from '#/pages/dashboard/components/KeyboardShortcut'
 import { useInputBindings } from '#/providers/InputBindingsProvider'
 import { unsafeEntries } from '#/utilities/object'
 import { useText } from '$/providers/react'
