+++ app/gui/src/dashboard/pages/dashboard/components/AssetIcon.tsx
