+++ app/gui/src/dashboard/pages/dashboard/components/AssetSummary.tsx
@@ -1,13 +1,13 @@
 /** @file Displays a few details of an asset. */
 import BreadcrumbArrowIcon from '#/assets/breadcrumb_arrow.svg'
-import AssetIcon from '#/components/dashboard/AssetIcon'
 import { Text } from '#/components/Text'
+import AssetIcon from '#/pages/dashboard/components/AssetIcon'
 import type * as backend from '#/services/Backend'
 import * as tailwindMerge from '#/utilities/tailwindMerge'
 import { useText } from '$/providers/react'
 import * as dateTime from 'enso-common/src/utilities/data/dateTime'
-import { Badge } from '../Badge'
-import { Icon } from '../Icon'
+import { Badge } from '../../../components/Badge'
+import { Icon } from '../../../components/Icon'
 
 /** Props for an {@link AssetSummary}. */
 export interface AssetSummaryProps {
