+++ app/gui/src/dashboard/pages/dashboard/components/FileNameColumn.tsx
@@ -1,8 +1,8 @@
 /** @file The icon and name of a {@link FileAsset}. */
-import type { AssetColumnProps } from '#/components/dashboard/column'
 import EditableSpan from '#/components/EditableSpan'
 import { Icon } from '#/components/Icon'
 import { useGetAssetChildren } from '#/layouts/Drive/assetsTableItemsHooks'
+import type { AssetColumnProps } from '#/pages/dashboard/components/column'
 import { titleSchema, type FileAsset } from '#/services/Backend'
 import { fileIcon } from '#/utilities/fileIcon'
 import { merger } from '#/utilities/object'
