+++ app/gui/src/dashboard/pages/dashboard/components/KeyboardShortcut.tsx
