+++ app/gui/src/dashboard/pages/dashboard/components/PermissionDisplay.tsx
