+++ app/gui/src/dashboard/pages/dashboard/components/ProjectIcon.tsx
