+++ app/gui/src/dashboard/pages/dashboard/components/column/LabelsColumn.tsx
@@ -3,14 +3,14 @@ import DotsIcon from '#/assets/dots.svg'
 import { Button } from '#/components/Button'
 import ContextMenu from '#/components/ContextMenu'
 import ContextMenuEntry from '#/components/ContextMenuEntry'
-import type * as column from '#/components/dashboard/column'
-import Label from '#/components/dashboard/Label'
 import { Dialog, Popover } from '#/components/Dialog'
 import { backendMutationOptions } from '#/hooks/backendHooks'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useMeasureCallback } from '#/hooks/measureHooks'
 import { useToastAndLog } from '#/hooks/toastAndLogHooks'
 import ManageLabelsModal from '#/modals/ManageLabelsModal'
+import type * as column from '#/pages/dashboard/components/column'
+import Label from '#/pages/dashboard/components/Label'
 import { setModal, unsetModal } from '#/providers/ModalProvider'
 import { FALLBACK_COLOR } from '#/services/Backend'
 import { mergeRefs } from '#/utilities/mergeRefs'
