+++ app/gui/src/dashboard/pages/dashboard/components/column/PlaceholderColumn.tsx
