+++ app/gui/src/dashboard/pages/dashboard/components/column/columnUtils.ts
