+++ app/gui/src/dashboard/pages/dashboard/components/columnHeading/ModifiedColumnHeading.tsx
@@ -2,9 +2,9 @@
 import { Button } from '#/components/Button'
 import { Icon } from '#/components/Icon'
 import { Text } from '#/components/Text'
-import type { AssetColumnHeadingProps } from '#/components/dashboard/column'
-import { Column } from '#/components/dashboard/column/columnUtils'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
+import type { AssetColumnHeadingProps } from '#/pages/dashboard/components/column'
+import { Column } from '#/pages/dashboard/components/column/columnUtils'
 import { SortDirection, iconIdFor, nextSortDirection } from '#/utilities/sorting'
 import { twJoin } from '#/utilities/tailwindMerge'
 import { useText } from '$/providers/react'
