+++ app/gui/src/dashboard/pages/dashboard/components/columnHeading/PathColumnHeading.tsx
